+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Improvements
 * simplify Dockerfile and remove docker build support for `LOGPREP_VERSION`
+* pytest.param now works with test_cases document generation
 
 ### Bugfix
 * generic resolver now follows yaml standard and accepts a list instead of relying on the ordering of a dict